DialogContent

+++ src/main/java/family/helpful/persist/repository/ChannelContentRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;
 
-public interface ChannelContentRepository extends JpaRepository<ChannelContent, Integer> {
+public interface ChannelContentRepository extends JpaRepository<ChannelContent, Long> {
     @Query("SELECT c FROM ChannelContent c WHERE c.channel.name = ?1 order by c.createDate DESC")
     List<ChannelContent> findByTitleWithAmount(String title, Pageable pageable);
 }
