+++ src/main/java/family/helpful/persist/repository/FamilyRepository.java
@@ -4,6 +4,6 @@ import family.helpful.persist.message.model.Family;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface FamilyRepository extends JpaRepository<Family, Integer> {
+public interface FamilyRepository extends JpaRepository<Family, Long> {
 
 }
