+++ src/main/java/family/helpful/persist/repository/ProblemContentRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;
 
-public interface ProblemContentRepository extends JpaRepository<ProblemContent, Integer> {
+public interface ProblemContentRepository extends JpaRepository<ProblemContent, Long> {
     @Query("SELECT c FROM ProblemContent c WHERE c.problemTitle.name = ?1 order by c.currentThankAmount DESC")
     List<ProblemContent> findByTitleWithAmount(String title, Pageable pageable);
 }
